requirements.txt
aiogram==3.0.0
beanie==1.22.6
python-dotenv==1.0.0